refactor(Login): refactored and added the requested changes
refactor(Login): refactored and added the requested changes

refactor(Login): added constant strings
