+++ src/components/Login/Login.tsx
@@ -1,74 +1,9 @@
 import React, { useReducer, useEffect } from 'react'
 import Card from '../UI/Card';
-
-
-type State = {
-  username: string
-  password: string
-  isButtonDisabled: boolean
-  helperText: string
-  isError: boolean
-}
-
-const initialState:State = {
-  username: '',
-  password: '',
-  isButtonDisabled: true,
-  helperText: '',
-  isError: false
-}
-
-type Action = { type: 'setUsername', payload: string  }
-  | { type: 'setPassword', payload: string }
-  | { type: 'setIsButtonDisabled', payload: boolean }
-  | { type: 'loginSuccess', payload: string }
-  | { type: 'loginFailed', payload: string }
-  | { type: 'setIsError', payload: boolean };
-
-
-const reducer = (state: State, action: Action): State => {
-  switch (action.type) {
-    case 'setUsername':
-      return {
-        ...state,
-        username: action.payload
-      };
-    case 'setPassword':
-      return {
-        ...state,
-        password: action.payload
-      };
-    case 'setIsButtonDisabled':
-      return {
-        ...state,
-        isButtonDisabled: action.payload
-      };
-    case 'loginSuccess':
-      return {
-        ...state,
-        helperText: action.payload,
-        isError: false
-      };
-    case 'loginFailed':
-      return {
-        ...state,
-        helperText: action.payload,
-        isError: true
-      };
-    case 'setIsError':
-      return {
-        ...state,
-        isError: action.payload
-    };
-  }
-}
-
-
-
-
+import reducer, { initialState } from './LoginReducer';
 
 const Login = () => {
-  const [state, dispatch] = useReducer(reducer, initialState);
+  const [state, dispatch] = useReducer(reducer, initialState)
 
   useEffect(() => {
     if (state.username.trim() && state.password.trim()) {
@@ -121,8 +56,13 @@ const Login = () => {
         });
       }
 
+
+      const submitFormHandler = (event: React.FormEvent) => {
+        event.preventDefault;
+      }
+
   return (
-    <form noValidate autoComplete="off">
+    <form noValidate autoComplete="off" onSubmit={submitFormHandler}>
       <Card class='card'>
         <div>
           <input id='username' type='email' placeholder='Username' onChange={usernameChangeHandler} onKeyPress={keyPressHandler}/>
