+++ src/utils/Constants.ts
@@ -1,0 +1,8 @@
+const constants = {
+  user: 'setUsername',
+  pass:  'setPassword',
+  success: 'loginSuccess',
+  fail: 'loginFailed'
+}
+
+export default constants;
